ENYO-2297: Creation of the application. Modification of the README.md.
app.js : Core of the application. Have the controllers link to it. The start of the application, how to fill the repeater.
controllers.js : Three controllers. One for the collection for the repeater. One for the change of uri when an entry of the repeater is selected to change the detail controller.
model.js : Two models. One for the collection controller and the other one for the model controller.
main.js : Main view of the application. Where the controllers are bound to UI components. One repeater and another cmponent for the detail.
When an entry is selected, the uri of the messageController is changed and it triggers the request and change of the detail controller.
We create a component for each attribute of the model with a transform for the binding to insure of the format.
OData.js : enyo.OData is used to make the request to the service.